requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
importlib-metadata==1.7.0
Markdown==3.2.1
Pillow==7.1.2
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
zipp==3.1.0
